+++ packages/abstract-document/package.json
@@ -31,10 +31,10 @@
     "@types/base64-js": "^1.2.5",
     "@types/blob-stream": "^0.1.30",
     "@types/isomorphic-fetch": "^0.0.34",
-    "@types/jszip": "0.0.31",
     "@types/pdfkit": "^0.7.31",
     "@types/ramda": "^0.25.29",
     "@types/react": "^16.9.0",
-    "@types/react-dom": "^16.9.0"
+    "@types/react-dom": "^16.9.0",
+    "jszip": "^3.7.1"
   }
 }
