+++ packages/abstract-document/src/_tests_/abstract-document-exporters/docx2/export-test-def.ts
@@ -1,0 +1,10 @@
+/**
+ * Defines a data-type that represents a single test for the Docx render function
+ */
+export type ExportTestDef = {
+  readonly name: string;
+  readonly only?: boolean;
+  readonly skip?: boolean;
+  readonly abstractDocJsx: JSX.Element;
+  readonly expectedDocxZipContexts: { readonly [filename: string]: string };
+};
