+++ packages/abstract-document/src/_tests_/abstract-document-exporters/docx2/test-defs/simple-test.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import { ExportTestDef } from "../export-test-def";
+import { Paragraph, AbstractDoc, Section } from "../../../../abstract-document-jsx";
+
+export const test: ExportTestDef = {
+  name: "simple test",
+  abstractDocJsx: (
+    <AbstractDoc>
+      <Section>
+        <Paragraph></Paragraph>
+      </Section>
+    </AbstractDoc>
+  ),
+  expectedDocxZipContexts: {
+    olle: "kalle",
+  },
+};
